langchainStream/requirements.txt
streamlit
langchain
langchain-openai
pandas
requests
python-dotenv
tabulate
